Implemented role based authorization Moved configs to one package
